tests/wpt/web-platform-tests/largest-contentful-paint/video-data-uri.html

<!doctype html>
<html>
<title>This test verifies a video element of data uri src triggers an LCP entry
  to be emitted</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <script>
    const get_lcp_entry = () => {
      return new Promise(resolve => {
        new PerformanceObserver((list, observer) => {
          if (list.getEntries()) {
            observer.disconnect();
            resolve(list.getEntries()[0]);
          }
        }).observe({ type: "largest-contentful-paint", buffered: true });
      });
    };

    promise_test(async t => {
      lcpEntry = await get_lcp_entry();
      assert_true(lcpEntry.url.startsWith('data:video/mp4;base64,T2dnUwAC'));
    }, "Video of data URI src should trigger an LCP entry to be emitted.")
  </script>
  <!--
    This is the base64 encoding of the images/pattern.ogv video file.
  -->
  <video autoplay muted src="data:video/mp4;base64,T2dnUwACAAAAAAAAAADmWaI+AAAAAKRLSUwBKoB0aGVvcmEDAgEAAgACAAAUAAAUAAwAAAAeAAAA
    AQAAAQAAAQAAAACw2E9nZ1MAAAAAAAAAAAAA5lmiPgEAAADp90reDj////////////////+QgXRo
    ZW9yYQ0AAABMYXZmNTcuMjYuMTAwAQAAAB8AAABlbmNvZGVyPUxhdmM1Ny4yNC4xMDUgbGlidGhl
    b3JhgnRoZW9yYb7NKPe5zWsYtalJShBznOYxjFKUpCEIMYxiEIQhCEAAAAAAAAAAAAARba5TZ5LI
    /FYS/Hg5W2zmKvVoq1QoEykkWhD+eTmbjWZTCXiyVSmTiSSCGQh8PB2OBqNBgLxWKhQJBGIhCHw8
    HAyGAsFAiDgVFtrlNnksj8VhL8eDlbbOYq9WirVCgTKSRaEP55OZuNZlMJeLJVKZOJJIIZCHw8HY
    4Go0GAvFYqFAkEYiEIfDwcDIYCwUCIOBQLDw8PDw8PDw8PDw8PDw8PDw8PDw8PDw8PDw8PDw8PDw
    8PDw8PDw8PDw8PDw8PDw8PDw8PDw8PDw8PDw8PDw8PDw8MDA8SFBQVDQ0OERIVFRQODg8SFBUVFQ
    4QERMUFRUVEBEUFRUVFRUSExQVFRUVFRQVFRUVFRUVFRUVFRUVFRUQDAsQFBkbHA0NDhIVHBwbDg
    0QFBkcHBwOEBMWGx0dHBETGRwcHh4dFBgbHB0eHh0bHB0dHh4eHh0dHR0eHh4dEAsKEBgoMz0MDA
    4TGjo8Nw4NEBgoOUU4DhEWHTNXUD4SFiU6RG1nTRgjN0BRaHFcMUBOV2d5eGVIXF9icGRnYxMTEx
    MTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTExMTEx
    MTExMSEhUZGhoaGhIUFhoaGhoaFRYZGhoaGhoZGhoaGhoaGhoaGhoaGhoaGhoaGhoaGhoaGhoaGh
    oaGhoaGhoaGhoaERIWHyQkJCQSFBgiJCQkJBYYISQkJCQkHyIkJCQkJCQkJCQkJCQkJCQkJCQkJC
    QkJCQkJCQkJCQkJCQkJCQkJBESGC9jY2NjEhUaQmNjY2MYGjhjY2NjYy9CY2NjY2NjY2NjY2NjY2
    NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2MVFRUVFRUVFRUVFRUVFRUVFRUVFRUVFRUVFRUVFRUVFR
    UVFRUVFRUVFRUVFRUVFRUVFRUVFRUVFRUVFRUVFRUVEhISFRcYGRsSEhUXGBkbHBIVFxgZGxwdFR
    cYGRscHR0XGBkbHB0dHRgZGxwdHR0eGRscHR0dHh4bHB0dHR4eHhERERQXGhwgEREUFxocICIRFB
    caHCAiJRQXGhwgIiUlFxocICIlJSUaHCAiJSUlKRwgIiUlJSkqICIlJSUpKioQEBAUGBwgKBAQFB
    gcICgwEBQYHCAoMEAUGBwgKDBAQBgcICgwQEBAHCAoMEBAQGAgKDBAQEBggCgwQEBAYICAB8Xlx0
    fV7c7D8vrrAaZid8hRvB1RN7csxFuo43wH7lEkS9wbGS+tVSNMyuxdiECcjB7R1Ml85htasNjKpS
    vPt3D8k7iGmZXYuxBC+RR4arUGxkvH5y7mJXR7R5Jwn3VUhBiuap91VIrsaCM5TSg9o867khwMrW
    Y2+cP4rwvBLzt/wnHaYe0edSRMYC6tZmU1BrvhktIUf2gXoU8bHMuyNA7lB7R51ym213sFcFKowI
    viT/i0Wscg+4RDubX+4haRsMxZWgN05K5FD3bzqS9VSVCPM4TpWs2C43ihFdgaSByeKHu3Xf/2TG
    8tgpB7PAtOs7jixWYw+Ayo5GjUTSybX/1KW52RxYfB8nBNLJtHgt4DPq6BZWBFpjyZX/1KW5Ca0e
    vOwG1EX/A9j5fQm5hOz6W2CtcCaWTXTFAeZO71VIgCTX69y9TiaXag3Os2ES1DcLKw0/xR5HfnCq
    kpQF0Z1kxKNfhZWLycml2keduHMQh3HubB/pbUUoCK5wxetZRZWPJF/bdyE21H2YjMOhP/pkthqK
    UCOEWVm68+1J5n7ahES5sOhaZPdOC5j4kc91FVIsrF8ofe+A2on/16Z4RiKQZcMU3NouO9N4YAvr
    WaiA6h4bfLqhTitbnnJ2iPSVRNJH+aZGE+YXzq7Ah/OncW2K59AKamlocOUYTSvaJPNcjDfMGrmG
    9pOV2MbgI9v3B3ECZ7RLJ51UpzMn0C1huA87Ngom9lkiaw3t5yvFZmDl1HpkuP+PiqlawgD69jAT
    5Nxr2i6cwiytcwHhK2KJvZI9C1m/4VUil8RvO/ydxmgsFdzdgGpMbUeyyRNOi1k5hMb6hVSMuTrO
    E/xuDhGExQ219l07sV2kG5fOEnkWHwgqUkbvC0P2KTytY4nHLqJDc3DMGlDbX2aXK/4UuJxizaIk
    ZITS7a3HN5374PrVlYKIcP9xl1BUKqQ7aAml2k1o5uGcN8A+tPz1HF1YVnmE7cyx4FIiUA2ml1k0
    hX9HB7l4tMO+R9YrMWcf5Anub1BZXUp3Ce4jBM21l0kyhcF/vg6FGeHa345MYv4BVSciTJhj5Abu
    D2K0dfIXc4jKAbazaS53rv1lYqpIVr2fcgcPox4u/WVnRfJ25GGING2s2cqjKIVUtwGbRtrljLd9
    CQOHhewUTfiKxWk7Olr2dHyIKlLgejEbasmmdGF/dhuhVrU9xGi6Hksgm/+5Bw813T3mJyRNqIYG
    dYspVZFzQ6dhNLJ7H+fYWh8Q+cMbzLc/O0evM4srXGjpECaXaT2jApqM4LRavgPnH7ecDRQSErab
    X3zC4EcXfOVZZUpYs3UIfMsKVR+6hgFzHhvWWWl4EqZtrJpHnyeO0T2icPrqVRyyDRKmbayexv7w
    dolGfh1hwtsK4G5jDOIHz/lTULUM47PaBmNJm2ssmTq+ssXeHBjgij3G5P+u5QVFIGQ21TNM5aGO
    HbqKssQ/HiM9kvcWjdCtF6gZNMzbXFhNP2gV2FNQi+OpOR+S+3RvOBVSOr+E5hjyPrQho7/QDNEG
    2qRNLpHl6WVl3m4p3POFvwEWUN0ByvCQTSttdM48H7tjQWVk73qoUvhiSDbVK0mzyohbuHXofmEa
    K/xXYJ+Vq7tBUN6lMAdrouC3p96IS8kMzbVK0myY4f+HKdRGsrG9SlDwEfQkXsGLIbapmmcv/sA5
    TrqC36t4sRdjylU4JC9KwG2plM0zxuT2iFFzAPXyj9ZWRu+tx5UpFv0jn0gQrKyMF5MyaZsDbXG7
    /qIdp0tHG4jOQumLzBliaZttaLfZFUBSOu7FaUn/+IXETfwUj2E0o6gJ2HB/l8N7jFnzWWBESEra
    bWPvy9bUKqS4y78CME0rbXSTNFRf8H7r1wwxQbltish5nFVIRkhKaTNtc6L3LHAh8+B2yi/tHvXG
    4nusVwAKMb/0/MCmoWrvASDM0mbay5YRI+7CtC96OPtxudDEyTGmbbWVRgkvR8qaiA8+rLCft7cW
    8H8UI3E8nzmJVSQIT3+0srHfUbgKA21ZNM8WEy+W7wbj9OuBpm21MKGWN80kaA5PZfoSqkRPLa1h
    31wIEjiUhcnX/e5VSWVkQnPhtqoYXrjLFpn7M8tjB17xSqfWgoA21StJpM48eSG+5A/dsGUQn8sV
    7impA4dQjxPyrsBfHd8tUGBIJWkxtrnljE3eu/xTUO/nVsA9I4uVlZ5uQvy9IwYjbWUmaZ5XE9HA
    WVkXUKmoI3y4vDKZpnKNtccJHK2iA83ej+fvgI3KR9P6qpG/kBCUdxHFisLkq8aZttTCZlj/b0G8
    XoLX/3fHhZWCVcMsWmZtqmYXz0cpOiBHCqpKUZu76iICRxYVuSULpmF/421MsWmfyhbP4ew1FVKA
    jFlY437JXImUTm2r/4ZYtMy61hf16RPJIRA8tU1BDc5/JzAkEzTM21lyx7sK9wojRX/OHXoOv05I
    DbUymaZyscL7qlMA8c/CiK3csceqzuOEU1EPpbz4QEahIShpm21MJmWN924f98WKyf51EEYBli0z
    NtUzC+6X9P9ysrU1CHyA3RJFFr1w67HpyULT+YMsWmZtquYXz97oKil44sI1bpL8hRSDeMkhiIBw
    OgxwZ5Fs6+5M+NdH+3Kjv0sreSqqRvGSQxEA4HQY4M8i2dfcmfGuj/blR36WVvJVVI3jJIYiAcDo
    McGeRbOvuTPjXR/tyo79LK3kqqkVUnCfqAES8EzTM21lykY4Q+LKxby+9F3ZHR/uC2OGpS9cv6BZ
    XAebhckMGIymaZm2st8/B38i6A/n58pVLKwfURet4UBwSF6UaZttSZljhd2jW9BZWcrX0/hG4Sdt
    /SBCdH6UMJmWK80zba3URKaik8iB9PR2459CuyOAbi0/GWLTMmYXm2t0vUkNQhRPVldKpAN5HgHy
    ZfdOtGuj/YxwZ5S8u3CjqMgQoyQJRdawvJlE530/+sVg21c8GWLTPf3yJVSVUoCMWVjjfslciZRO
    bav/hli0zLrWF/XpE8khT2dnUwAAQAAAAAAAAADmWaI+AgAAAA6lVqcC/yoswcgPXVDgwIQQYBK8
    JHGwwhNHk7ANjkOfhcKbwVnSYLxZeR6bjpWsA1Y6R+wZcHLxjMJmal1gZAIcaRD6tAh8GgJhGSjo
    RpMND9CVoehCgT8+hDwjEFfgrNOcG5ocU5t9Rh4kcJd9R5dXmYNMXqo9IPXZaceU+4mxu3IDOM1n
    B9miEvZlRdJVqYnBiCi1YTjCdVTR09SWSpdB0VbJ7X5K4yFui7Z1Jd5eyf7J8/JTx+2ji97qbeSC
    43FWTwEqySHBdRDnj5dGfPoc5Z1yIDxOmzKKkkTEqU7vDvwgj0xCs82zZqaO10ruMxgqZ/lTOZZm
    VXKlRKlXdtEJ3duzrZ6ubMGIlwLV6+zHVrI8HdzncTtng2MaugcSSKFo5x4tqwQh7fGnvW140wBP
    Z2dTAABLAAAAAAAAAOZZoj4DAAAAwc+ctgsAAAAAAAAAAAAAAE9nZ1MAAEADAAAAAAAA5lmiPgQA
    AADHLitMAv8qLMHID11Q4MCEEGASvCRxsMITR5OwDY5Dn4XCm8FZ0mC8WXkem46VrANWOkfsGXBy
    8YzCZmpdYGQCHGkQ+rQIfBoCYRko6EaTDQ/QlaHoQoE/PoQ8IxBX4KzTnBuaHFObfUYeJHCXfUeX
    V5mDTF6qPSD12WnHlPuJsbtyAzjNZwfZohL2ZUXSVamJwYgotWE4wnVU0dPUlkqXQdFWye1+SuMh
    bou2dSXeXsn+yfPyU8fto4ve6m3kguNxVk8BKskhwXUQ54+XRnz6HOWdciA8TpsyipJExKlO7w78
    II9MQrPNs2amjtdK7jMYKmf5UzmWZlVypUSpV3bRCd3bs62ermzBiJcC1evsx1ayPB3c53E7Z4Nj
    GroHEkihaOceLasEIe3xp71teNMAT2dnUwAASwMAAAAAAADmWaI+BQAAAL37AWwLAAAAAAAAAAAA
    AABPZ2dTAABABgAAAAAAAOZZoj4GAAAA5hY3pwL/KizByA9dUODAhBBgErwkcbDCE0eTsA2OQ5+F
    wpvBWdJgvFl5HpuOlawDVjpH7BlwcvGMwmZqXWBkAhxpEPq0CHwaAmEZKOhGkw0P0JWh6EKBPz6E
    PCMQV+Cs05wbmhxTm31GHiRwl31Hl1eZg0xeqj0g9dlpx5T7ibG7cgM4zWcH2aIS9mVF0lWpicGI
    KLVhOMJ1VNHT1JZKl0HRVsntfkrjIW6LtnUl3l7J/snz8lPH7aOL3upt5ILjcVZPASrJIcF1EOeP
    l0Z8+hzlnXIgPE6bMoqSRMSpTu8O/CCPTEKzzbNmpo7XSu4zGCpn+VM5lmZVcqVEqVd20Qnd27Ot
    nq5swYiXAtXr7MdWsjwd3OdxO2eDYxq6BxJIoWjnHi2rBCHt8ae9bXjTAE9nZ1MAAEsGAAAAAAAA
    5lmiPgcAAACJ2bKjCwAAAAAAAAAAAAAAT2dnUwAAQAkAAAAAAADmWaI+CAAAAP/6iUwC/yoswcgP
    XVDgwIQQYBK8JHGwwhNHk7ANjkOfhcKbwVnSYLxZeR6bjpWsA1Y6R+wZcHLxjMJmal1gZAIcaRD6
    tAh8GgJhGSjoRpMND9CVoehCgT8+hDwjEFfgrNOcG5ocU5t9Rh4kcJd9R5dXmYNMXqo9IPXZaceU
    +4mxu3IDOM1nB9miEvZlRdJVqYnBiCi1YTjCdVTR09SWSpdB0VbJ7X5K4yFui7Z1Jd5eyf7J8/JT
    x+2ji97qbeSC43FWTwEqySHBdRDnj5dGfPoc5Z1yIDxOmzKKkkTEqU7vDvwgj0xCs82zZqaO10ru
    MxgqZ/lTOZZmVXKlRKlXdtEJ3duzrZ6ubMGIlwLV6+zHVrI8HdzncTtng2MaugcSSKFo5x4tqwQh
    7fGnvW140wBPZ2dTAABLCQAAAAAAAOZZoj4JAAAAZcFzUwsAAAAAAAAAAAAAAE9nZ1MAAEAMAAAA
    AAAA5lmiPgoAAADewpWnAv8qLMHID11Q4MCEEGASvCRxsMITR5OwDY5Dn4XCm8FZ0mC8WXkem46V
    rANWOkfsGXBy8YzCZmpdYGQCHGkQ+rQIfBoCYRko6EaTDQ/QlaHoQoE/PoQ8IxBX4KzTnBuaHFOb
    fUYeJHCXfUeXV5mDTF6qPSD12WnHlPuJsbtyAzjNZwfZohL2ZUXSVamJwYgotWE4wnVU0dPUlkqX
    QdFWye1+SuMhbou2dSXeXsn+yfPyU8fto4ve6m3kguNxVk8BKskhwXUQ54+XRnz6HOWdciA8Tpsy
    ipJExKlO7w78II9MQrPNs2amjtdK7jMYKmf5UzmWZlVypUSpV3bRCd3bs62ermzBiJcC1evsx1ay
    PB3c53E7Z4NjGroHEkihaOceLasEIe3xp71teNMAT2dnUwAASwwAAAAAAADmWaI+CwAAAFHjwJwL
    AAAAAAAAAAAAAABPZ2dTAABADwAAAAAAAOZZoj4MAAAAF0noTAL/KizByA9dUODAhBBgErwkcbDC
    E0eTsA2OQ5+FwpvBWdJgvFl5HpuOlawDVjpH7BlwcvGMwmZqXWBkAhxpEPq0CHwaAmEZKOhGkw0P
    0JWh6EKBPz6EPCMQV+Cs05wbmhxTm31GHiRwl31Hl1eZg0xeqj0g9dlpx5T7ibG7cgM4zWcH2aIS
    9mVF0lWpicGIKLVhOMJ1VNHT1JZKl0HRVsntfkrjIW6LtnUl3l7J/snz8lPH7aOL3upt5ILjcVZP
    ASrJIcF1EOePl0Z8+hzlnXIgPE6bMoqSRMSpTu8O/CCPTEKzzbNmpo7XSu4zGCpn+VM5lmZVcqVE
    qVd20Qnd27Otnq5swYiXAtXr7MdWsjwd3OdxO2eDYxq6BxJIoWjnHi2rBCHt8ae9bXjTAE9nZ1MA
    AEsPAAAAAAAA5lmiPg0AAAAt111GCwAAAAAAAAAAAAAAT2dnUwAAQBIAAAAAAADmWaI+DgAAALvQ
    BAYC/yoswcgPXVDgwIQQYBK8JHGwwhNHk7ANjkOfhcKbwVnSYLxZeR6bjpWsA1Y6R+wZcHLxjMJm
    al1gZAIcaRD6tAh8GgJhGSjoRpMND9CVoehCgT8+hDwjEFfgrNOcG5ocU5t9Rh4kcJd9R5dXmYNM
    Xqo9IPXZaceU+4mxu3IDOM1nB9miEvZlRdJVqYnBiCi1YTjCdVTR09SWSpdB0VbJ7X5K4yFui7Z1
    Jd5eyf7J8/JTx+2ji97qbeSC43FWTwEqySHBdRDnj5dGfPoc5Z1yIDxOmzKKkkTEqU7vDvwgj0xC
    s82zZqaO10ruMxgqZ/lTOZZmVXKlRKlXdtEJ3duzrZ6ubMGIlwLV6+zHVrI8HdzncTtng2MaugcS
    SKFo5x4tqwQh7fGnvW140wBPZ2dTAABLEgAAAAAAAOZZoj4PAAAAgHc9kAsAAAAAAAAAAAAAAE9n
    Z1MAAEAVAAAAAAAA5lmiPhAAAAD0zWMtAv8qLMHID11Q4MCEEGASvCRxsMITR5OwDY5Dn4XCm8FZ
    0mC8WXkem46VrANWOkfsGXBy8YzCZmpdYGQCHGkQ+rQIfBoCYRko6EaTDQ/QlaHoQoE/PoQ8IxBX
    4KzTnBuaHFObfUYeJHCXfUeXV5mDTF6qPSD12WnHlPuJsbtyAzjNZwfZohL2ZUXSVamJwYgotWE4
    wnVU0dPUlkqXQdFWye1+SuMhbou2dSXeXsn+yfPyU8fto4ve6m3kguNxVk8BKskhwXUQ54+XRnz6
    HOWdciA8TpsyipJExKlO7w78II9MQrPNs2amjtdK7jMYKmf5UzmWZlVypUSpV3bRCd3bs62ermzB
    iJcC1evsx1ayPB3c53E7Z4NjGroHEkihaOceLasEIe3xp71teNMAT2dnUwAASxUAAAAAAADmWaI+
    EQAAAKLKJiULAAAAAAAAAAAAAABPZ2dTAARAGAAAAAAAAOZZoj4SAAAAmbxPPwL/KizByA9dUODA
    hBBgErwkcbDCE0eTsA2OQ5+FwpvBWdJgvFl5HpuOlawDVjpH7BlwcvGMwmZqXWBkAhxpEPq0CHwa
    AmEZKOhGkw0P0JWh6EKBPz6EPCMQV+Cs05wbmhxTm31GHiRwl31Hl1eZg0xeqj0g9dlpx5T7ibG7
    cgM4zWcH2aIS9mVF0lWpicGIKLVhOMJ1VNHT1JZKl0HRVsntfkrjIW6LtnUl3l7J/snz8lPH7aOL
    3upt5ILjcVZPASrJIcF1EOePl0Z8+hzlnXIgPE6bMoqSRMSpTu8O/CCPTEKzzbNmpo7XSu4zGCpn
    +VM5lmZVcqVEqVd20Qnd27Otnq5swYiXAtXr7MdWsjwd3OdxO2eDYxq6BxJIoWjnHi2rBCHt8ae9
    bXjTAA==" width="600"></video>
</body>

</html>
